+++ lib/myWidgets/rapports.dart
@@ -1,27 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-// class myGridView extends StatelessWidget{
-//   @override
-//   Widget build(BuildContext context) {
-//     return GridView.count(
-//   primary: false,
-//   padding: const EdgeInsets.all(20),
-//   crossAxisSpacing: 10,
-//   mainAxisSpacing: 10,
-//   crossAxisCount: 2,
-//   children: listRapport,
-//        );
-//   }
-// }
-
-// Widget list(){
-//   for(var i=0; i < 12; i++){
-
-//   }
-//   return ;
-// }
-
 class Rapports extends StatelessWidget {
   const Rapports({super.key, this.text = "12 mars 2024 - 12 avril 2024", this.site = "KM4"});
   final String text;
@@ -154,49 +133,3 @@ class Rapports extends StatelessWidget {
 
 
 
-class Banniere extends StatelessWidget {
-  const Banniere({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-
-
-    double screenWidth = MediaQuery.of(context).size.width;
-    double screenHeight = MediaQuery.of(context).size.height;
-
-    // Calculate the responsive width and height
-
-    double containerHeight = screenHeight * 0.15; // 15% of screen height
-
-    return Column(
-      children: [
-        Container(
-          height: containerHeight,
-          width: double.infinity,
-          padding: EdgeInsets.all(8.0),
-          color: Color.fromARGB(255, 244, 244, 244),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-            Text("Mes Rapports", 
-            style: TextStyle(
-              fontFamily: GoogleFonts.outfit().fontFamily,
-              fontSize: 20, 
-              fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(
-                width: 262.79,
-                height: 27,
-                child: Text("Visualisez et téléchargez votre rapport de consommation électrique au format PDF,Excel ou CSV.",
-                style: TextStyle(
-                  fontSize: 10
-                ),))
-          ],
-          ),
-        ),
-      ],
-    );
-  }
-}
